i32 -> u16

+++ src/domain/aggregate/member.rs
@@ -6,14 +6,14 @@ use super::value_object::{grade::Grade, major::Major};
 pub struct Member {
     pub id: MemberId, // メンバーのID (Value Object)
     pub name: String,
-    pub age: i32,
+    pub age: u16,
     pub grade: Grade,
     pub major: Major,
 }
 
 impl Member {
     // メンバーの新規作成メソッド
-    pub fn new(name: String, age: i32, grade: Grade, major: Major) -> Self {
+    pub fn new(name: String, age: u16, grade: Grade, major: Major) -> Self {
         Member {
             id: MemberId::gen(),
             name,
@@ -24,7 +24,7 @@ impl Member {
     }
 
     // メンバーの再構成メソッド
-    pub fn reconstruct(id: MemberId, name: String, age: i32, grade: Grade, major: Major) -> Self {
+    pub fn reconstruct(id: MemberId, name: String, age: u16, grade: Grade, major: Major) -> Self {
         Member {
             id,
             name,
